Split node and epoch functions.
